Move home link position
Moved the home link from the top of the page to the bottom of the
recipe card. I did this because it was recommended to have the link home
at the bottom of the page instead of the top, so that it is easier to
return to the home page after scrolling through the recipe. I saw this
recommendation on another person's recipes project in the creations
showcase section in the odin project's discord.